Move back build pages to airing today module
